Add session tokens with cookies
And uncover some really weird behaviour where if the cookie is set on
`/account/login` through a fetch request with `credentials: "include"`,
it will be sent in any future requests to `/account/login` and
`/account/signup` but NOT to `/arguments` (where I need it). So I had to
change the endpoints to `/login`, `/signup` and `/arguments` (which I
actually don't mind and probably would have done at some point anyway).
But the facinating thing is I didn't find any information on this
anywhere but I just felt like this was the problem and so when I tried
renaming the end points and everything worked which it was bizarre.